multi-model-prd/pyproject.toml
[project]
name = "multi-model-prd"
version = "0.1.0"
description = "Multi-model collaborative discussion PRD generation system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "litellm>=1.40.0",
    "langgraph>=0.0.20",
    "redis>=5.0.0",
    "chromadb>=0.4.0",
    "structlog>=24.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "weasyprint>=60.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
